# perf: make contracts load faster