--- a/src/ape/contracts/__init__.py
+++ b/src/ape/contracts/__init__.py
@@ -1,4 +1,8 @@
-from .base import ContractContainer, ContractEvent, ContractInstance, ContractLog, ContractNamespace
+def __getattr__(name: str):
+    import ape.contracts.base as module
+
+    return getattr(module, name)
+
 
 __all__ = [
     "ContractContainer",