### What I did

before:

```
time python -c "import ape.contracts.base"
python -c "import ape.contracts.base"  2.48s user 0.61s system 143% cpu 2.145 total
```

after:

```
python -c "import ape.contracts.base"  1.17s user 0.17s system 99% cpu 1.338 total
```

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated