@antazoey 2024-11-02
> Eventually would love a refactor of this entire module, it's far too condensed into .base

yes, smaller modules seem better for python